backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Vector store & embedding
faiss-cpu>=1.7.4
numpy>=1.24.0

# Document parsing
pypdf>=4.0.0
python-docx>=1.1.0

# Alibaba Bailian (DashScope)
openai>=1.12.0
dashscope>=1.14.0

# LangChain & agents
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
